Change the freq of timer

+++ device/timer.c
@@ -1,7 +1,6 @@
 #include "timer.h"
 #include "io.h"
 #include "print.h"
-#include <stdint.h>
 
 #define IRQ0_FREQUENCY	   100
 #define INPUT_FREQUENCY	   1193180
@@ -12,7 +11,7 @@
 #define READ_WRITE_LATCH   3
 #define PIT_CONTROL_PORT   0x43
 
-static void frequency_set(uint8_t counter_port, uint8_t counter_no, uint8_t rwl, uint8_t counter_mode, uint8_t counter_value) {
+static void frequency_set(uint8_t counter_port, uint8_t counter_no, uint8_t rwl, uint8_t counter_mode, uint16_t counter_value) {
   outb(PIT_CONTROL_PORT, (uint8_t)(counter_no << 6 | rwl << 4 | counter_mode << 1));
   outb(counter_port, (uint8_t)counter_value);
   outb(counter_port, (uint8_t)counter_value >> 8);
@@ -20,6 +19,6 @@ static void frequency_set(uint8_t counter_port, uint8_t counter_no, uint8_t rwl,
 
 void timer_init(){
   put_str("timer_init start\n");
-  frequency_set(COUNTER0_PORT, COUNTER0_NO, READ_WRITE_LATCH, COUNTER_MODE, (uint8_t)COUNTER0_VALUE);
+  frequency_set(COUNTER0_PORT, COUNTER0_NO, READ_WRITE_LATCH, COUNTER_MODE, (uint16_t)COUNTER0_VALUE);
   put_str("timer_init done\n");
 }
